Change page selector to use tabs instead of card
