+++ client/src/components/dashboard/TemplateSelector.js
@@ -3,7 +3,7 @@
  * to select a theme in Researchify dashboard page.
  */
 import React, { useState } from 'react';
-import { Container, Button, Modal, Form, Col, Image } from 'react-bootstrap';
+import { Container, Button, Form, Col, Image } from 'react-bootstrap';
 import { useDispatch } from 'react-redux';
 import { updateTeamTheme } from '../../actions/team';
 import toast from 'react-hot-toast';
@@ -82,114 +82,101 @@ const TemplateSelector = (props) => {
   };
 
   return (
-    <Modal
-      show={props.displayModal}
-      onHide={props.closeModal}
-      centered
-      size="lg"
+    <Form
+      className="researchify-github-form"
+      noValidate
+      onSubmit={handleSubmit}
     >
-      <Modal.Header closeButton>
-        <Modal.Title id="contained-modal-title-vcenter">
-          Select a colour combination and layout.
-        </Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
-        <Form
-          className="researchify-github-form"
-          noValidate
-          onSubmit={handleSubmit}
-        >
-          <Form.Group controlId="theme">
-            <Form.Label>Select a Theme Colour</Form.Label>
-            <Container fluid>
-              <Form.Row>
-                <Form.Check
-                  inline
-                  type="radio"
-                  name="theme"
-                  value="theme1"
-                  onChange={updateForm}
-                  checked={theme === 1}
-                />
-                <div className="theme-icon theme-1-icon"></div>
-                <Form.Check
-                  inline
-                  type="radio"
-                  name="theme"
-                  value="theme2"
-                  onChange={updateForm}
-                  checked={theme === 2}
-                />
-                <div className="theme-icon theme-2-icon"></div>
-                <Form.Check
-                  inline
-                  type="radio"
-                  name="theme"
-                  value="theme3"
-                  onChange={updateForm}
-                  checked={theme === 3}
-                />
-                <div className="theme-icon theme-3-icon"></div>
-              </Form.Row>
-            </Container>
-            <Form.Control.Feedback type="invalid">
-              Please select a theme.
-            </Form.Control.Feedback>
-          </Form.Group>
+      <Form.Group controlId="theme">
+        <Form.Label>Select a Theme Colour</Form.Label>
+        <Container fluid>
+          <Form.Row>
+            <Form.Check
+              inline
+              type="radio"
+              name="theme"
+              value="theme1"
+              onChange={updateForm}
+              checked={theme === 1}
+            />
+            <div className="theme-icon theme-1-icon"></div>
+            <Form.Check
+              inline
+              type="radio"
+              name="theme"
+              value="theme2"
+              onChange={updateForm}
+              checked={theme === 2}
+            />
+            <div className="theme-icon theme-2-icon"></div>
+            <Form.Check
+              inline
+              type="radio"
+              name="theme"
+              value="theme3"
+              onChange={updateForm}
+              checked={theme === 3}
+            />
+            <div className="theme-icon theme-3-icon"></div>
+          </Form.Row>
+        </Container>
 
-          <Form.Group controlId="layout">
-            <Form.Label>Select a Layout</Form.Label>
-            <Container fluid>
-              <Form.Row>
-                <Col className="layout-display">
-                  <Form.Check
-                    checked={formInputs.layout === 1}
-                    inline
-                    type="radio"
-                    name="layout"
-                    label="Layout 1"
-                    value={1}
-                    className="form-radio-text"
-                    onChange={updateForm}
-                  />
-                  <Image src={singleColumnLayout} className="img-fluid" />
-                </Col>
-                <Col className="layout-display">
-                  <Form.Check
-                    checked={formInputs.layout === 2}
-                    inline
-                    type="radio"
-                    name="layout"
-                    label="Layout 2"
-                    value={2}
-                    className="form-radio-text"
-                    onChange={updateForm}
-                  />
-                  <Image src={fShapeLayout} className="img-fluid" />
-                </Col>
-                <Col className="layout-display">
-                  <Form.Check
-                    checked={formInputs.layout === 3}
-                    inline
-                    type="radio"
-                    name="layout"
-                    label="Layout 3"
-                    value={3}
-                    className="form-radio-text"
-                    onChange={updateForm}
-                  />
-                  <Image src={zigZagLayout} className="img-fluid" />
-                </Col>
-              </Form.Row>
-            </Container>
-          </Form.Group>
+        <Form.Control.Feedback type="invalid">
+          Please select a theme.
+        </Form.Control.Feedback>
+      </Form.Group>
 
-          <Button id="submitButton" type="submit">
-            Create Website
-          </Button>
-        </Form>
-      </Modal.Body>
-    </Modal>
+      <Form.Group controlId="layout">
+        <Form.Label>Select a Layout</Form.Label>
+        <Container fluid>
+          <Form.Row>
+            <Col className="layout-display">
+              <Form.Check
+                checked={formInputs.layout === 1}
+                inline
+                type="radio"
+                name="layout"
+                label="Layout 1"
+                value={1}
+                className="form-radio-text"
+                onChange={updateForm}
+              />
+              <Image src={singleColumnLayout} className="img-fluid" />
+            </Col>
+            <Col className="layout-display">
+              <Form.Check
+                checked={formInputs.layout === 2}
+                inline
+                type="radio"
+                name="layout"
+                label="Layout 2"
+                value={2}
+                className="form-radio-text"
+                onChange={updateForm}
+              />
+              <Image src={fShapeLayout} className="img-fluid" />
+            </Col>
+            <Col className="layout-display">
+              <Form.Check
+                checked={formInputs.layout === 3}
+                inline
+                type="radio"
+                name="layout"
+                label="Layout 3"
+                value={3}
+                className="form-radio-text"
+                onChange={updateForm}
+              />
+              <Image src={zigZagLayout} className="img-fluid" />
+            </Col>
+          </Form.Row>
+        </Container>
+      </Form.Group>
+
+        <Button id="submitButton" type="submit">
+          Create Website
+        </Button>
+    </Form>
   );
 };
